+++ packages/core/src/sdk/index.js
@@ -1,0 +1,1 @@
+export { SDKRegistry, UnknownChainError } from './SDKRegistry.js';